+++ ORBITER/src/lib/foundations/graphics/graphics.h
@@ -414,7 +414,7 @@ public:
 
 
 	void draw_graph(int x, int y,
-			int dx, int dy, int nb_V, int *Edges, int nb_E);
+			int dx, int dy, int nb_V, long int *Edges, int nb_E);
 	void draw_graph_with_distinguished_edge(
 		int x, int y,
 		int dx, int dy, int nb_V, int *Edges, int nb_E,
@@ -426,7 +426,7 @@ public:
 			int *Edges, int nb_E, int *coords_2D, int *Base,
 			int f_point_labels, int point_label_offset, int f_directed);
 	void draw_tournament(int x, int y,
-			int dx, int dy, int nb_V, int *Edges, int nb_E,
+			int dx, int dy, int nb_V, long int *Edges, int nb_E,
 			int verbose_level);
 	void draw_bitmatrix2(int f_dots,
 		int f_partition, int nb_row_parts, int *row_part_first,
